新增用户增加性别和生日 -> 演示MyBatis/Jackson使用枚举

+++ src/main/java/com/qnxy/blog/mapper/UserInfoMapper.java
@@ -15,7 +15,7 @@ public interface UserInfoMapper {
      */
     UserInfo selectByUsernameAndPassword(String username, String password);
 
-    UserInfo selectByPrimaryKey(Integer id);
+    UserInfo selectByPrimaryKey(Long id);
 
     /**
      * 根据用户名称查询用户信息
